Move mempoolIDs back to mempool/ids.go (#467)
